style: rename ShuffleArray to Shuffle
