+++ src/Lotto.js
@@ -11,7 +11,6 @@ class Lotto	{
 		if (numbers.length !== 6) {
 			throw new Error("[ERROR] 로또 번호는 6개여야 합니다.");
 		}
-
 	}
 
 	getNumberArray() {
